Move ExtractedSecrets to suites
