+++ rustls/src/lib.rs
@@ -336,8 +336,6 @@ mod check;
 mod bs_debug;
 mod builder;
 mod enums;
-#[cfg(feature = "extract_secrets")]
-mod extract_secrets;
 mod key;
 mod key_log;
 mod key_log_file;
@@ -369,8 +367,6 @@ pub use crate::conn::{
 };
 pub use crate::enums::{CipherSuite, ProtocolVersion, SignatureScheme};
 pub use crate::error::Error;
-#[cfg(feature = "extract_secrets")]
-pub use crate::extract_secrets::{AlgorithmSecrets, ExtractedSecrets};
 pub use crate::key::{Certificate, PrivateKey};
 pub use crate::key_log::{KeyLog, NoKeyLog};
 pub use crate::key_log_file::KeyLogFile;
@@ -380,6 +376,8 @@ pub use crate::msgs::enums::{
 };
 pub use crate::msgs::handshake::{DigitallySignedStruct, DistinguishedNames};
 pub use crate::stream::{Stream, StreamOwned};
+#[cfg(feature = "extract_secrets")]
+pub use crate::suites::{AlgorithmSecrets, ExtractedSecrets};
 pub use crate::suites::{
     BulkAlgorithm, SupportedCipherSuite, ALL_CIPHER_SUITES, DEFAULT_CIPHER_SUITES,
 };
